+++ client/src/modules/fiscal/templates/modals/fiscal.closing.modal.js
@@ -9,7 +9,10 @@ ClosingFYModalCtrl.$inject = [
 ];
 
 // The closing fiscal year controller
-function ClosingFYModalCtrl(Notify, Fiscal, Modal, Session, Instance, Data, uiGridConstants, bhConstants, Tree, Accounts) {
+function ClosingFYModalCtrl(
+  Notify, Fiscal, Modal, Session, Instance, Data,
+  uiGridConstants, bhConstants, Tree, Accounts
+) {
   const vm = this;
   const fiscalYearId = Data.id;
